fifo complete
finished fifo and added a function that will print out the data in a clean manner.

+++ process.h
@@ -4,6 +4,6 @@ typedef struct process {
 	int burst;
 	int finishTime;
 	int turnAroundTime;
-	int normalizedTurnAround;
+	float normalizedTurnAround;
 	
 }Process;
